ref(ui): move pages to own folders
