+++ frontend/src/components/App.tsx
@@ -17,28 +17,26 @@ import {
 } from "react-router-dom"
 
 import { Container, ContainerBase } from "@/components/Base"
-import { CurrentKeys } from "@/components/CurrentKeys"
 import ErrorBoundary from "@/components/ErrorBoundary"
 import { Helmet } from "@/components/Helmet"
-import HelpMenuModal from "@/components/HelpMenuModal"
-import Login from "@/components/Login"
-import NoMatch from "@/components/NoMatch"
-import PasswordResetConfirmation from "@/components/PasswordResetConfirmation"
-import Recipe from "@/components/Recipe"
-import Recipes from "@/components/RecipeList"
-import Schedule from "@/components/Schedule"
-import TeamCreate from "@/components/TeamCreate"
-import TeamInvite from "@/components/TeamInvite"
-import Teams from "@/components/Teams"
-import AddRecipe from "@/containers/AddRecipe"
-import Home from "@/containers/Home"
 import Notification from "@/containers/Notification"
-import PasswordChange from "@/containers/PasswordChange"
-import PasswordReset from "@/containers/PasswordReset"
-import PasswordSet from "@/containers/PasswordSet"
-import Settings from "@/containers/Settings"
-import Signup from "@/containers/Signup"
-import Team from "@/containers/Team"
+import NotFound from "@/pages/404/404.page"
+import Home from "@/pages/index/Index.page"
+import Login from "@/pages/login/Login.page"
+import PasswordChangePage from "@/pages/password-change/PasswordChange.page"
+import PasswordResetPage from "@/pages/password-reset/PasswordReset.page"
+import PasswordResetConfirmPage from "@/pages/password-reset-confirm/PasswordResetConfirm.page"
+import PasswordSet from "@/pages/password-set/PasswordSet.page"
+import RecipeCreatePage from "@/pages/recipe-create/RecipeCreate.page"
+import RecipeDetailPage from "@/pages/recipe-detail/RecipeDetail.page"
+import RecipeListPage from "@/pages/recipe-list/RecipeList.page"
+import SchedulePage from "@/pages/schedule/Schedule.page"
+import SettingsPage from "@/pages/settings/Settings.page"
+import SignupPage from "@/pages/signup/Signup.page"
+import TeamCreatePage from "@/pages/team-create/TeamCreate.page"
+import TeamDetailPage from "@/pages/team-detail/TeamDetail.page"
+import TeamInvitePage from "@/pages/team-invite/TeamInvite.page"
+import TeamsListPage from "@/pages/team-list/TeamList.page"
 import store, { history, IState } from "@/store/store"
 import { theme, ThemeProvider } from "@/theme"
 
@@ -145,16 +143,19 @@ function Base() {
           <DndProvider backend={HTML5Backend}>
             <ErrorBoundary>
               <Helmet />
-              <CurrentKeys />
               <ConnectedRouter history={history}>
                 <ScrollRestore />
                 <Switch>
                   <PublicOnlyRoute exact path="/login" component={Login} />
-                  <PublicOnlyRoute exact path="/signup" component={Signup} />
+                  <PublicOnlyRoute
+                    exact
+                    path="/signup"
+                    component={SignupPage}
+                  />
                   <Route
                     exact
                     path="/password-reset"
-                    component={PasswordReset}
+                    component={PasswordResetPage}
                   />
                   <ContainerBase>
                     <Switch>
@@ -162,44 +163,44 @@ function Base() {
                       <PrivateRoute
                         exact
                         path="/schedule/:type(shopping|recipes)?"
-                        component={Schedule}
+                        component={SchedulePage}
                       />
                       <PrivateRoute
                         exact
                         path="/t/:id(\d+)(.*)/schedule/:type(shopping|recipes)?"
-                        component={Schedule}
+                        component={SchedulePage}
                       />
                       <Container>
                         <Switch>
                           <Route
                             exact
                             path="/password-reset/confirm/:uid([0-9A-Za-z_\-]+).:token([0-9A-Za-z]{1,13}-[0-9A-Za-z]{1,20})"
-                            component={PasswordResetConfirmation}
+                            component={PasswordResetConfirmPage}
                           />
                           <PrivateRoute
                             exact
                             path="/recipes/add"
-                            component={AddRecipe}
+                            component={RecipeCreatePage}
                           />
                           <PrivateRoute
                             exact
                             path="/recipes/"
-                            component={Recipes}
+                            component={RecipeListPage}
                           />
                           <PrivateRoute
                             exact
                             path="/recipes/:id(\d+)(.*)"
-                            component={Recipe}
+                            component={RecipeDetailPage}
                           />
                           <PrivateRoute
                             exact
                             path="/settings"
-                            component={Settings}
+                            component={SettingsPage}
                           />
                           <PrivateRoute
                             exact
                             path="/password"
-                            component={PasswordChange}
+                            component={PasswordChangePage}
                           />
                           <PrivateRoute
                             exact
@@ -209,25 +210,29 @@ function Base() {
                           <PrivateRoute
                             exact
                             path="/t/create"
-                            component={TeamCreate}
+                            component={TeamCreatePage}
                           />
                           <PrivateRoute
                             exact
                             path="/t/:id(\d+)(.*)/invite"
-                            component={TeamInvite}
+                            component={TeamInvitePage}
                           />
                           <PrivateRoute
                             exact
                             path="/t/:id(\d+)(.*)/settings"
-                            component={Team}
+                            component={TeamDetailPage}
                           />
                           <PrivateRoute
                             exact
                             path="/t/:id(\d+)(.*)"
-                            component={Team}
+                            component={TeamDetailPage}
                           />
-                          <PrivateRoute exact path="/t/" component={Teams} />
-                          <Route component={NoMatch} />
+                          <PrivateRoute
+                            exact
+                            path="/t/"
+                            component={TeamsListPage}
+                          />
+                          <Route component={NotFound} />
                         </Switch>
                       </Container>
                     </Switch>
@@ -235,7 +240,6 @@ function Base() {
                 </Switch>
               </ConnectedRouter>
               <Notification />
-              <HelpMenuModal />
             </ErrorBoundary>
           </DndProvider>
         </HelmetProvider>
