--- frontend/src/components/NoMatch.tsx
@@ -1,8 +1,0 @@
-const NoMatch = () => (
-  <section>
-    <h1 className="fs-10rem bold text-center">404</h1>
-    <p className="fs-3rem text-center">Nothing here 🌵</p>
-  </section>
-)
-
-export default NoMatch
